+++ include/nexuralnet/dnn/layers/computational_layers/max_pooling_layer.h
@@ -36,7 +36,7 @@ namespace nexural {
 
 		}
 
-		virtual void Setup(const LayerShape& prevLayerShape, const int layerIndex) {
+		virtual void Setup(const LayerShape& prevLayerShape, const size_t layerIndex) {
 			if ((prevLayerShape.GetNR() % _kernel_height != 0) || (prevLayerShape.GetNC() % _kernel_width != 0)) {
 				throw std::runtime_error("Cannot apply max pooling to the input layer!");
 			}
