+++ include/nexuralnet/dnn/layers/computational_layers/tanh_layerh.h
@@ -35,7 +35,7 @@ namespace nexural {
 
 		}
 
-		virtual void Setup(const LayerShape& prevLayerShape, const int layerIndex) {
+		virtual void Setup(const LayerShape& prevLayerShape, const size_t layerIndex) {
 			_inputShape.Resize(prevLayerShape);
 			_outputShape.Resize(_inputShape);
 			_outputData.Resize(_outputShape);
